// JavaScriptProjects/ReactProjects/warehouse_wars/front-end/ww-react-ts/src/types/game/ClientSend.ts
export interface PlayerReady extends GamePackage {
    type: "player_ready";
    player_id: string;
    room_id: string;
}
//export function isPlayerReady (object: any): object is PlayerReady {
//    return object && object.type === "player_ready" 
//        && typeof object.player_id === 'string' 
//        && typeof object.room_id === 'string';
//}



export interface PlayerUnready extends GamePackage {
    type: "player_unready";
    player_id: string;
    room_id: string;
}
//export function isPlayerUnready (object: any): object is PlayerReady {
//    return object && object.type === "player_unready" 
//        && typeof object.player_id === 'string' 
//        && typeof object.room_id === 'string';
//}



export interface PlayerMovement extends GamePackage {
    type: "player_movement";
    player_id: string;
    room_id: string;
    dir: [number,number];
    isPulling: boolean;
    updateNum: number;
}
// ...
export interface RequestUpdate extends GamePackage {
    type: "request_update";
    player_id: string;
    room_id: string;
}